update embedding example
