+++ examples/flutter_embedding/lib/components/flutter_app_container.dart
