+++ examples/flutter_embedding/lib/components/ripple_loader.dart
