+++ examples/flutter_embedding/lib/widgets/app.dart
