+++ examples/flutter_embedding/pubspec.yaml
@@ -8,6 +8,7 @@ environment:
 
 dependencies:
   cupertino_icons: 1.0.5
+  firebase_core: ^2.15.0
   flutter:
     sdk: flutter
   flutter_riverpod: ^2.1.3
@@ -18,11 +19,11 @@ dependencies:
 
 dev_dependencies:
   build_runner: ^2.4.0
-  build_web_compilers: ^4.0.0
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.0
   jaspr_builder: ^0.4.0
+  jaspr_web_compilers: ^4.0.4
 
 flutter:
   uses-material-design: true
