+++ examples/flutter_embedding/web/main_jaspr.dart
@@ -1,9 +1,7 @@
 import 'package:element_embedding_demo/components/app.dart';
-import 'package:element_embedding_demo/interop/flutter.dart';
 import 'package:jaspr/browser.dart';
 
 /// The web entrypoint for the jaspr app.
 void main() {
   runApp(App());
-  runFlutterApp(attachTo: '#flutter_target');
 }
