--- experiments/flutter_embedding_v2/lib/main_flutter.dart
@@ -1,2 +1,0 @@
-/// This is currently needed as a flutter compilation target.
-void main() {}
