--- experiments/flutter_embedding_v2/lib/providers/effects_provider.dart
@@ -1,4 +1,0 @@
-import 'package:riverpod/riverpod.dart';
-
-final effectsProvider = StateProvider((ref) => <String>{});
-final rotationProvider = StateProvider((ref) => 0.0);
