--- experiments/flutter_embedding_v2/pubspec.yaml
@@ -1,30 +1,0 @@
-name: element_embedding_demo_v2
-description: A small app to be embedded into a HTML element
-version: 1.0.0
-publish_to: none
-
-environment:
-  sdk: '>=3.0.0-0 <4.0.0'
-
-dependencies:
-  cupertino_icons: 1.0.5
-  flutter:
-    sdk: flutter
-  flutter_riverpod: ^2.1.3
-  jaspr: ^0.4.0
-  jaspr_riverpod: ^0.2.0
-  js: ^0.6.7
-  riverpod: ^2.1.3
-
-dev_dependencies:
-  build_runner: ^2.4.0
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^2.0.0
-  jaspr_builder: ^0.4.0
-  jaspr_web_compilers: ^4.0.4
-
-flutter:
-  uses-material-design: true
-  assets:
-    - assets/dash.png
