--- experiments/flutter_embedding_v2/web/assets/AssetManifest.json
@@ -1,1 +1,0 @@
-{"assets/dash.png":["assets/dash.png"],"packages/cupertino_icons/assets/CupertinoIcons.ttf":["packages/cupertino_icons/assets/CupertinoIcons.ttf"]}
